pyproject.toml
[project]
name = "multi-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "google-genai>=1.25.0",
    "pillow>=11.3.0",
    "pipwin>=0.5.2",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.1.1",
    "speechrecognition>=3.14.3",
]
